♻ refactor: Temporary refactor project to JS
