+++ index.html
@@ -5,13 +5,13 @@
     <link
       rel="icon"
       type="image/svg+xml"
-      href="/src/asstets/images/favicon.svg"
+      href="/src/assets/images/favicon.svg"
     />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Bookstore React Redux</title>
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
