+++ src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
-import { store } from './redux/store/store';
+import { store } from './redux/store';
 import './style/main.scss';
 
 ReactDOM.render(
