+++ src/redux/books/actionCreators.js
@@ -1,13 +1,13 @@
 import { ADD_BOOK, REMOVE_BOOK } from './actionTypes';
 
-export const addBook = (book: Book) => {
+export const addBook = book => {
   return {
     type: ADD_BOOK,
     payload: book
   };
 };
 
-export const removeBook = (book: Book) => {
+export const removeBook = book => {
   return {
     type: REMOVE_BOOK,
     payload: book
