+++ src/redux/categories/categoriesReducer.js
@@ -1,0 +1,26 @@
+const initialState = [
+  {
+    id: 1,
+    category: 'Fiction'
+  },
+  {
+    id: 2,
+    category: 'Non-Fiction'
+  },
+  {
+    id: 3,
+    category: 'Mystery'
+  },
+  {
+    id: 4,
+    category: 'Horror'
+  },
+  {
+    id: 5,
+    category: 'Romance'
+  }
+];
+
+export const categoriesReducer = (state = initialState) => {
+  return state;
+};
