--- src/redux/categories/categoriesReducer.ts
@@ -1,20 +1,0 @@
-const initialState: CategoriesState = {
-  categories: [
-    {
-      id: 1,
-      title: 'Category 1'
-    },
-    {
-      id: 2,
-      title: 'Category 2'
-    },
-    {
-      id: 3,
-      title: 'Category 3'
-    }
-  ]
-};
-
-export const categoriesReducer = (state: CategoriesState = initialState) => {
-  return state;
-};
