+++ src/redux/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import { booksReducer } from '../books/booksReducer';
-import { categoriesReducer } from '../categories/categoriesReducer';
+import { booksReducer } from './books/booksReducer';
+import { categoriesReducer } from './categories/categoriesReducer';
 
 export const rootReducer = combineReducers({
   books: booksReducer,
